feat: ensure user is created with password hashed
