+++ package.json
@@ -13,6 +13,7 @@
   "license": "ISC",
   "devDependencies": {
     "@rocketseat/eslint-config": "2.2.2",
+    "@types/bcryptjs": "2.4.6",
     "@types/node": "22.14.1",
     "eslint": "8.57.1",
     "prisma": "6.6.0",
@@ -22,6 +23,7 @@
   },
   "dependencies": {
     "@prisma/client": "6.6.0",
+    "bcryptjs": "3.0.2",
     "dotenv": "16.5.0",
     "fastify": "5.3.2",
     "zod": "3.24.3"
