+++ src/http/controllers/register.controller.ts
@@ -1,0 +1,33 @@
+import { prisma } from '@/lib/prisma'
+import { hash } from 'bcryptjs'
+import { FastifyReply, FastifyRequest } from 'fastify'
+import { z } from 'zod'
+
+export async function register(request: FastifyRequest, reply: FastifyReply) {
+  const registerBodySchema = z.object({
+    name: z.string(),
+    email: z.string().email(),
+    password: z.string().min(6),
+  })
+
+  const { email, name, password } = registerBodySchema.parse(request.body)
+
+  const SALT_ROUNDS = 6
+  const passwordHash = await hash(password, SALT_ROUNDS)
+
+  const userWithSameEmail = await prisma.user.findUnique({ where: { email } })
+
+  if (userWithSameEmail) {
+    return reply.status(409).send()
+  }
+
+  await prisma.user.create({
+    data: {
+      name,
+      email,
+      password_hash: passwordHash,
+    },
+  })
+
+  return reply.status(201).send()
+}
