+++ src/http/routes.ts
@@ -1,0 +1,6 @@
+import { FastifyInstance } from 'fastify'
+import { register } from './controllers/register.controller'
+
+export async function appRoutes(app: FastifyInstance) {
+  app.post('/users', register)
+}
